NodeLoaderWorker must be contained inside WorkerSupport otherwise LoaderWorker cannot be found
